java task 1

+++ java_support_frameworks/src/main/java/com/epam/support_farmeworks/food/EatCharacteristic.java
@@ -8,7 +8,7 @@ import lombok.Getter;
 @Getter
 @Builder(builderClassName = "FoodCollector", builderMethodName = "createCharacteristic")
 public class EatCharacteristic {
-    private float eaten;
+    private double eaten;
     private DayPart dayPart;
     private CatPosition catPosition;
 }
